SnatchAlert/SnatchAlert/requirements.txt
Django==5.2.8
djangorestframework==3.15.2
djangorestframework-simplejwt==5.4.0
django-filter==24.3
django-cors-headers==4.6.0
drf-yasg==1.21.8
psycopg2-binary==2.9.10
Pillow==11.0.0
python-decouple==3.8
